Let exactly one breakpoint type claim a feature line
Found by the first item on the debugger checklist: every feature line, on both
majors, drew a second and then a third breakpoint circle. Two bugs stacked.

Karate 2 detection for breakpoints was project-wide (LibraryUtil.getLibraryRoots),
so in a monorepo with one v2 module - testProjects/karate-versions is exactly
that - the Karate breakpoint type claimed the v1 module's feature files too. The
run configuration already had module-scoped detection with the reasoning worked
out; that logic moves to KarateLibraries, and the breakpoint type resolves the
module from the file.

And on a genuine v2 file, both the Java line breakpoint type (through
debugger.javaDebugAware) and the Karate one accepted the line, which the platform
renders as inline breakpoint variants - a choice with one right answer. A Java
breakpoint could never bind on v2 anyway, since there is no per-step Java method,
so KarateDebugAware now steps aside for v2 files and keeps v1 exactly as it was.
